Make ENABLE_GPU OFF code work
