+++ cpu_framework/include/dataset.h
@@ -29,7 +29,9 @@ public:
     {
         static_assert(DIMS > 0, "This interface is not intended for ZEROD variables");
         static_assert(DIMS + 1 == sizeof...(Indices), "You have provided the wrong number of arguments");
+#ifdef ENABLE_GPU
         in_const_operator = false;
+#endif
         return indexer<T, DIMS>::access(static_cast<T*>(m_data), m_offsets, DIMS+1, static_cast<Indices&&>(idx)...);
     }
 
@@ -38,7 +40,9 @@ public:
     {
         static_assert(DIMS > 0, "This interface is not intended for ZEROD variables");
         static_assert(DIMS + 1 == sizeof...(Indices), "You have provided the wrong number of arguments");
+#ifdef ENABLE_GPU
         in_const_operator = true;
+#endif
         return indexer<T, DIMS>::access_const(static_cast<T*>(m_data), m_offsets, DIMS+1, static_cast<Indices&&>(idx)...);
     }
