+++ gpu_framework/include/gpu_globals.h
@@ -18,13 +18,6 @@ class GPUManager_t;
 }
 CDF_GLOBAL GDF::GPUManager_t* gpu_manager;
 
-// Global variable to store the system page size (generally 4KB/ 4096B)
-CDF_GLOBAL int system_page_size;
-
-// Ordered set which stores the m_data value and it's corresponding DSB pointer in order to look up the DSB which triggered the pagefault
-class dataSetBase;
-CDF_GLOBAL std::set<std::pair<void*, dataSetBase*>> dsb_addr_set;
-
 #ifdef GPU_MEM_LOG
 CDF_GLOBAL std::map<void*, strict_fp_t> gpu_mem_map;
 CDF_GLOBAL strict_fp_t tot_gpu_mem_used;
